requirements.txt
matplotlib
pandas
requests
arabic_reshaper
python-bidi